completely revamped the slider. Works great, except the editor hates it
