+++ src/components/Slider/Slider.tsx
@@ -1,0 +1,116 @@
+import * as React from 'react';
+
+import { styled } from 'styletron-react';
+import { ThemeProps } from '../theme';
+
+import { CanBeVertical, CanBeSelected } from './';
+import { SliderBar } from './SliderBar';
+import { Actions, resizeOnPointerMove } from './Resizer';
+
+export interface SliderProps extends ThemeProps {
+  isVertical: boolean;
+  // takes two children, which must be flex items or flex containers (if they contain resizable components)
+  // children: [SliderItem | SliderContainer, SliderItem | SliderContainer],
+  children: [JSX.Element, JSX.Element] ;
+  // the minimum number of pixels each component can be
+  minSizes: [number, number],
+  // the initial proportional split of the elements
+  sizes: [number, number],
+}
+
+interface SliderItemProps extends CanBeVertical {
+  $flexGrow: number,
+}
+const SliderContainer = styled('div', (props: CanBeVertical & CanBeSelected) => ({
+  display: 'flex',
+  overflow: 'auto',
+  alignItems: 'stretch',
+  width: '100%',
+  height: '100%',
+  flexDirection: (props.$vertical) ? 'row' : 'column',
+  cursor: props.selected ? (props.$vertical ? 'col-resize' : 'row-resize') : null,
+  // border: '3px solid red',
+}));
+const SliderItem = styled('div', (props: SliderItemProps) => ({
+  display: 'flex',
+  overflow: 'auto',
+  flexGrow: props.$flexGrow,
+  // border: '5px solid blue',
+}));
+
+export const Slider = function (props: SliderProps) {
+  const { isVertical, theme, children, minSizes, sizes } = props;
+
+  const itemRef0 = React.useRef<HTMLDivElement>(null);
+  const itemRef1 = React.useRef<HTMLDivElement>(null);
+
+  const [state, dispatch] = React.useReducer(resizeOnPointerMove, {
+    isVertical: isVertical,
+    grows: sizes,
+    refs: [itemRef0, itemRef1],
+    minSizes: minSizes,
+    resizing: false,
+  });
+
+  const [selected, setSelected] = React.useState(false);
+
+  const onMouseMove = (e: MouseEvent) => {
+    dispatch({ actionType: Actions.MouseMove, x: e.pageX, y: e.pageY });
+  };
+  const onMouseUp = (e: MouseEvent) => {
+    setSelected(false);
+    dispatch({ actionType: Actions.MouseUp, x: e.pageX, y: e.pageY });
+
+    window.removeEventListener('mousemove', onMouseMove);
+    window.removeEventListener('mouseup', onMouseUp);
+  };
+  const onMouseDown = (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
+    setSelected(true);
+    dispatch({ actionType: Actions.MouseDown, x: e.pageX, y: e.pageY });
+
+    window.addEventListener('mousemove', onMouseMove);
+    window.addEventListener('mouseup', onMouseUp);
+  };
+
+  // TODO: make sure we support only 1 touch more explicitly
+  const onTouchMove = (e: TouchEvent) => {
+    // only support single touch events
+    if (e.touches.length > 1) return;
+
+    dispatch({ actionType: Actions.MouseMove, x: e.touches[0].pageX, y: e.touches[0].pageY });
+  };
+  const onTouchEnd = (e: TouchEvent) => {
+    dispatch({ actionType: Actions.MouseUp, x: null, y: null });
+
+    window.removeEventListener('touchmove', onTouchMove);
+    window.removeEventListener('touchend', onTouchEnd);
+    window.removeEventListener('touchcancel', onTouchEnd);
+  };
+  const onTouchStart = (e: React.TouchEvent<HTMLDivElement>) => {
+    // only support single touch events
+    if (e.touches.length > 1) return;
+
+    dispatch({ actionType: Actions.MouseDown, x: e.touches[0].pageX, y: e.touches[0].pageY });
+
+    window.addEventListener('touchmove', onTouchMove);
+    window.addEventListener('touchend', onTouchEnd);
+    window.addEventListener('touchcancel', onTouchEnd);
+  };
+
+  // console.log('size: ', width, height, state.size);
+
+  return <SliderContainer $vertical={isVertical} selected={selected}>
+    <SliderItem ref={itemRef0} $flexGrow={state.grows[0]}>
+      {children[0]}
+    </SliderItem>
+
+    <SliderBar $vertical={isVertical} theme={theme} selected={state.resizing}
+      onMouseDownCallback={onMouseDown}
+      onTouchStartCallback={onTouchStart}
+    />
+
+    <SliderItem ref={itemRef1} $flexGrow={state.grows[1]}>
+      {children[1]}
+    </SliderItem>
+  </SliderContainer>;
+};
