+++ src/components/Slider/index.ts
@@ -1,0 +1,8 @@
+export { Slider, SliderProps } from './Slider';
+
+export interface CanBeVertical {
+  $vertical?: boolean
+}
+export interface CanBeSelected {
+  selected: boolean
+}
